requirements.txt
app-version==0.2.1
appdirs==1.4.3
asn1crypto==0.24.0
cached-property==1.3.1
cffi==1.11.5
characteristic==14.3.0
colorama==0.3.3
configparser==3.5.0
cryptography==2.4.2
cssselect==0.9.1
defusedxml==0.5.0
Django==1.11
django-ajax-selects==1.5.0
django-appconf==1.0.2
django-bootstrap3==11.0.0
django-chartit2==0.2.2
django-chosen==0.1
django-crispy-forms==1.5.0
django-debug-toolbar==1.3.2
django-dowser==0.1.4
django-easy-pdf==0.1.0
django-extensions==1.6.1
django-extra-views==0.7.1
django-firebird==1.8
django-fm==0.2.5
django-localflavor==1.5.2
django-material==0.8.0
django-session-security==2.6.0
django-suit==0.2.18
django-widget-tweaks==1.4.1
djangosecure==0.0.11
docopt==0.6.2
Dozer==0.5
enum34==1.1.4
fdb==1.6
firebirdsql==0.9.12
fpdf==1.7.2
future==0.16.0
GRR-M2Crypto==0.22.6.post2
html5lib==0.999999
httplib2==0.9.2
idna==2.1
ipaddress==1.0.16
isodate==0.5.4
lxml==4.2.5
Mako==1.0.3
MarkupSafe==0.23
mysqlclient==1.3.7
Pillow==2.9.0
pipreqs==0.4.2
pyasn1==0.1.9
pycparser==2.14
pycryptodome==3.6.1
PyJWT==1.0.0
PyPDF2==1.25
Pyphen==0.9.1
PySimpleSOAP==1.16
python-dateutil==2.5.3
python-slugify==1.2.0
pytz==2017.2
reportlab==3.2.0
requests==2.10.0
requests-toolbelt==0.8.0
simplejson==3.13.2
six==1.10.0
soap2py==1.16
SOAPpy==0.12.22
sqlparse==0.1.19
suds-jurko==0.6
tinycss==0.3
tolerance==0.1.2
Unidecode==0.4.19
virtualenv==13.1.0
WebOb==1.5.1
wstools==0.4.5
xhtml2pdf==0.0.6
xlrd==0.9.4
xlutils==1.7.1
xlwt==1.0.0
yarg==0.1.9
zeep==2.4.0
